requirements.txt
pandas>=2.0.0
requests>=2.25.0
pyyaml>=6.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0
sphinx-autodoc-typehints>=1.25.0 